requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.29
botocore==1.38.29
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
easydict==1.13
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gast==0.3.3
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.4
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
lightning==2.5.1.post0
lightning-cloud==0.5.70
lightning-utilities==0.14.3
lightning_sdk==0.2.18
litdata==0.2.45
litserve==0.2.11
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.55
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
paddleocr==3.0.2
paddlepaddle-gpu==3.0.0
paddlex==3.0.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
pi_heif==0.22.0
pikepdf==9.9.0
pillow==11.2.1
platformdirs==4.3.8
PPOCRLabel==3.1.0
premailer==3.10.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.6.1
pypdfium2==4.30.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.3
shapely==2.1.1
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.0+cu128
torchmetrics==1.3.1
torchvision==0.22.0+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trash-cli==0.24.5.26
triton==3.3.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
unstructured-client==0.37.1
unstructured.pytesseract==0.3.15
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
visualdl==2.5.3
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.0
